refactor(events): define orchestration SSE frames once
The producer and consumer each declared their own copy of
OrchestrationStreamEvent, and they disagreed: plan_created did not
parse, and iteration_complete bound to synthesizing. aura-events now
holds the single definition, matching what the server emits, and the
aura copy re-exports it. RoutingMode and IterationTimings move with it.

Emitted frames are unchanged; only the parsing side is corrected.

Ref: #623
